change folder structure, add utils
