+++ api/user.py
@@ -1,0 +1,8 @@
+from flask_restful import Resource
+
+
+class CreateUser(Resource):
+    def post(self):
+        return {
+            'status': 'success'
+        }
